Create group.services to group manager
